Cobot_Panda/calc/Code_source.py
# cobot_safety.py

import time

class CobotSafetyMonitor:
    def __init__(self):
        self.monitor_environment_called = False
        self.monitor_proximity_called = False
        self.monitor_force_torque_called = False
        self.monitor_vision_called = False
        self.emergency_stop_called = False
        self.detect_collision_called = False
        self.detect_collision_result = False
    
    def monitor_environment(self):
        self.monitor_environment_called = True
        # Your implementation here

    def monitor_proximity(self):
        self.monitor_proximity_called = True
        # Your implementation here
    
    def monitor_force_torque(self):
        self.monitor_force_torque_called = True
        # Your implementation here
    
    def monitor_vision(self):
        self.monitor_vision_called = True
        # Your implementation here
    
    def emergency_stop(self):
        self.emergency_stop_called = True
        # Your implementation here
    
    def detect_collision(self):
        self.detect_collision_called = True
        # Your implementation here
        return self.detect_collision_result
    
    def run(self):
        while True:
            self.monitor_environment()
            self.monitor_proximity()
            self.monitor_force_torque()
            self.monitor_vision()
            if self.detect_collision():
                self.emergency_stop()
            time.sleep(0.1)
